refactored currency form formatters
