+++ app/forms/mappings/Mappings.scala
@@ -24,6 +24,10 @@ import java.time.LocalDate
 
 trait Mappings extends Formatters with Constraints {
 
+  protected def currency(emptyErrorMessageKey: String, invalidErrorMessageKey: String): FieldMapping[BigDecimal] =
+    of(currencyFormatter(emptyErrorMessageKey, invalidErrorMessageKey))
+
+
   protected def text(errorKey: String = "error.required", args: Seq[String] = Seq.empty): FieldMapping[String] =
     of(stringFormatter(errorKey, args))
